Make gui into a folder
